refactor: разделить код на более мелкие переиспользуемые компоненты
