--- src/shared/ui/PhotoSection/PhotoSection.tsx
@@ -1,1005 +1,0 @@
-import clsx from 'clsx';
-import { Link } from 'react-router-dom';
-import { LikeButton } from '../LikeButton/LikeButton';
-import s from './PhotoSection.module.css';
-
-interface PhotoInfo {
-  id: string;
-  slug: string;
-  created_at: string;
-  updated_at: string;
-  promoted_at: string;
-  width: number;
-  height: number;
-  color: string;
-  blur_hash: string;
-  description: string | null;
-  alt_description: string;
-  breadcrumbs: [];
-  urls: {
-    raw: string;
-    full: string;
-    regular: string;
-    small: string;
-    thumb: string;
-    small_s3: string;
-  };
-  links: {
-    self: string;
-    html: string;
-    download: string;
-    download_location: string;
-  };
-  likes: number;
-  liked_by_user: boolean;
-  current_user_collections: [];
-  sponsorship: object | null;
-  topic_submissions: object;
-  user: object;
-}
-
-// Временное решение приходящих фотографий
-const photoList: PhotoInfo[] = [
-  {
-    id: 'seX13AzLqls',
-    slug: 'a-person-sitting-inside-of-a-tent-in-the-desert-seX13AzLqls',
-    created_at: '2023-04-28T13:32:49Z',
-    updated_at: '2024-01-28T05:49:49Z',
-    promoted_at: '2023-05-06T10:45:50Z',
-    width: 4500,
-    height: 3000,
-    color: '#262626',
-    blur_hash: 'L35hGRWq0MWBOtaz-njZENjF%0kC',
-    description:
-      'Nature Reserve - NEOM, Saudi Arabia | The NEOM Nature Reserve region is being designed to deliver protection and restoration of biodiversity across 95% of NEOM.',
-    alt_description: 'a person sitting inside of a tent in the desert',
-    breadcrumbs: [],
-    urls: {
-      raw: 'https://images.unsplash.com/photo-1682688759157-57988e10ffa8?ixid=M3w1NTQ1MTF8MXwxfGFsbHwxfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3',
-      full: 'https://images.unsplash.com/photo-1682688759157-57988e10ffa8?crop=entropy&cs=srgb&fm=jpg&ixid=M3w1NTQ1MTF8MXwxfGFsbHwxfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=85',
-      regular:
-        'https://images.unsplash.com/photo-1682688759157-57988e10ffa8?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MXwxfGFsbHwxfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=1080',
-      small:
-        'https://images.unsplash.com/photo-1682688759157-57988e10ffa8?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MXwxfGFsbHwxfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=400',
-      thumb:
-        'https://images.unsplash.com/photo-1682688759157-57988e10ffa8?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MXwxfGFsbHwxfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=200',
-      small_s3:
-        'https://s3.us-west-2.amazonaws.com/images.unsplash.com/small/photo-1682688759157-57988e10ffa8'
-    },
-    links: {
-      self: 'https://api.unsplash.com/photos/a-person-sitting-inside-of-a-tent-in-the-desert-seX13AzLqls',
-      html: 'https://unsplash.com/photos/a-person-sitting-inside-of-a-tent-in-the-desert-seX13AzLqls',
-      download:
-        'https://unsplash.com/photos/seX13AzLqls/download?ixid=M3w1NTQ1MTF8MXwxfGFsbHwxfHx8fHx8Mnx8MTcwNjQ1OTYxOXw',
-      download_location:
-        'https://api.unsplash.com/photos/seX13AzLqls/download?ixid=M3w1NTQ1MTF8MXwxfGFsbHwxfHx8fHx8Mnx8MTcwNjQ1OTYxOXw'
-    },
-    likes: 344,
-    liked_by_user: false,
-    current_user_collections: [],
-    sponsorship: {
-      impression_urls: [
-        'https://secure.insightexpressai.com/adServer/adServerESI.aspx?script=false&bannerID=11515602&rnd=[timestamp]&redir=https://secure.insightexpressai.com/adserver/1pixel.gif',
-        'https://secure.insightexpressai.com/adServer/adServerESI.aspx?script=false&bannerID=11515805&rnd=[timestamp]&redir=https://secure.insightexpressai.com/adserver/1pixel.gif'
-      ],
-      tagline: 'Made to Change',
-      tagline_url:
-        'https://www.neom.com/en-us?utm_source=unsplash&utm_medium=referral',
-      sponsor: {
-        id: 'mYizSrdJkkU',
-        updated_at: '2024-01-27T21:36:45Z',
-        username: 'neom',
-        name: 'NEOM',
-        first_name: 'NEOM',
-        last_name: null,
-        twitter_username: 'neom',
-        portfolio_url: 'http://www.neom.com',
-        bio: 'Located in the northwest of Saudi Arabia, NEOM’s diverse climate offers both sun-soaked beaches and snow-capped mountains. NEOM’s unique location will provide residents with enhanced livability while protecting 95% of the natural landscape.',
-        location: 'NEOM, Saudi Arabia',
-        links: {
-          self: 'https://api.unsplash.com/users/neom',
-          html: 'https://unsplash.com/@neom',
-          photos: 'https://api.unsplash.com/users/neom/photos',
-          likes: 'https://api.unsplash.com/users/neom/likes',
-          portfolio: 'https://api.unsplash.com/users/neom/portfolio',
-          following: 'https://api.unsplash.com/users/neom/following',
-          followers: 'https://api.unsplash.com/users/neom/followers'
-        },
-        profile_image: {
-          small:
-            'https://images.unsplash.com/profile-1679489218992-ebe823c797dfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=32&h=32',
-          medium:
-            'https://images.unsplash.com/profile-1679489218992-ebe823c797dfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=64&h=64',
-          large:
-            'https://images.unsplash.com/profile-1679489218992-ebe823c797dfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=128&h=128'
-        },
-        instagram_username: 'discoverneom',
-        total_collections: 7,
-        total_likes: 1,
-        total_photos: 202,
-        total_promoted_photos: 72,
-        accepted_tos: true,
-        for_hire: false,
-        social: {
-          instagram_username: 'discoverneom',
-          portfolio_url: 'http://www.neom.com',
-          twitter_username: 'neom',
-          paypal_email: null
-        }
-      }
-    },
-    topic_submissions: {},
-    user: {
-      id: 'mYizSrdJkkU',
-      updated_at: '2024-01-27T21:36:45Z',
-      username: 'neom',
-      name: 'NEOM',
-      first_name: 'NEOM',
-      last_name: null,
-      twitter_username: 'neom',
-      portfolio_url: 'http://www.neom.com',
-      bio: 'Located in the northwest of Saudi Arabia, NEOM’s diverse climate offers both sun-soaked beaches and snow-capped mountains. NEOM’s unique location will provide residents with enhanced livability while protecting 95% of the natural landscape.',
-      location: 'NEOM, Saudi Arabia',
-      links: {
-        self: 'https://api.unsplash.com/users/neom',
-        html: 'https://unsplash.com/@neom',
-        photos: 'https://api.unsplash.com/users/neom/photos',
-        likes: 'https://api.unsplash.com/users/neom/likes',
-        portfolio: 'https://api.unsplash.com/users/neom/portfolio',
-        following: 'https://api.unsplash.com/users/neom/following',
-        followers: 'https://api.unsplash.com/users/neom/followers'
-      },
-      profile_image: {
-        small:
-          'https://images.unsplash.com/profile-1679489218992-ebe823c797dfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=32&h=32',
-        medium:
-          'https://images.unsplash.com/profile-1679489218992-ebe823c797dfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=64&h=64',
-        large:
-          'https://images.unsplash.com/profile-1679489218992-ebe823c797dfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=128&h=128'
-      },
-      instagram_username: 'discoverneom',
-      total_collections: 7,
-      total_likes: 1,
-      total_photos: 202,
-      total_promoted_photos: 72,
-      accepted_tos: true,
-      for_hire: false,
-      social: {
-        instagram_username: 'discoverneom',
-        portfolio_url: 'http://www.neom.com',
-        twitter_username: 'neom',
-        paypal_email: null
-      }
-    }
-  },
-  {
-    id: 'WYmNFpyKn1o',
-    slug: 'a-restaurant-with-tables-chairs-and-plants-WYmNFpyKn1o',
-    created_at: '2024-01-27T13:39:50Z',
-    updated_at: '2024-01-28T08:44:15Z',
-    promoted_at: '2024-01-28T08:44:15Z',
-    width: 5623,
-    height: 3749,
-    color: '#26260c',
-    blur_hash: 'LDCP|jE000%g^*IVD*nhs;oeWBtR',
-    description: null,
-    alt_description: 'a restaurant with tables, chairs, and plants',
-    breadcrumbs: [],
-    urls: {
-      raw: 'https://images.unsplash.com/photo-1706362723628-60e8f1929ffe?ixid=M3w1NTQ1MTF8MHwxfGFsbHwyfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3',
-      full: 'https://images.unsplash.com/photo-1706362723628-60e8f1929ffe?crop=entropy&cs=srgb&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHwyfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=85',
-      regular:
-        'https://images.unsplash.com/photo-1706362723628-60e8f1929ffe?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHwyfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=1080',
-      small:
-        'https://images.unsplash.com/photo-1706362723628-60e8f1929ffe?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHwyfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=400',
-      thumb:
-        'https://images.unsplash.com/photo-1706362723628-60e8f1929ffe?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHwyfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=200',
-      small_s3:
-        'https://s3.us-west-2.amazonaws.com/images.unsplash.com/small/photo-1706362723628-60e8f1929ffe'
-    },
-    links: {
-      self: 'https://api.unsplash.com/photos/a-restaurant-with-tables-chairs-and-plants-WYmNFpyKn1o',
-      html: 'https://unsplash.com/photos/a-restaurant-with-tables-chairs-and-plants-WYmNFpyKn1o',
-      download:
-        'https://unsplash.com/photos/WYmNFpyKn1o/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHwyfHx8fHx8Mnx8MTcwNjQ1OTYxOXw',
-      download_location:
-        'https://api.unsplash.com/photos/WYmNFpyKn1o/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHwyfHx8fHx8Mnx8MTcwNjQ1OTYxOXw'
-    },
-    likes: 18,
-    liked_by_user: false,
-    current_user_collections: [],
-    sponsorship: null,
-    topic_submissions: {},
-    user: {
-      id: 'BdLxrRuTVwY',
-      updated_at: '2024-01-28T08:47:08Z',
-      username: 'elmartinbaron',
-      name: 'Martin Baron',
-      first_name: 'Martin',
-      last_name: 'Baron',
-      twitter_username: null,
-      portfolio_url: 'http://martinbaron.fr',
-      bio: 'Taking pictures for others to enjoy. If you like my work feel free to follow me on Instagram :-)',
-      location: 'Saumur',
-      links: {
-        self: 'https://api.unsplash.com/users/elmartinbaron',
-        html: 'https://unsplash.com/@elmartinbaron',
-        photos: 'https://api.unsplash.com/users/elmartinbaron/photos',
-        likes: 'https://api.unsplash.com/users/elmartinbaron/likes',
-        portfolio: 'https://api.unsplash.com/users/elmartinbaron/portfolio',
-        following: 'https://api.unsplash.com/users/elmartinbaron/following',
-        followers: 'https://api.unsplash.com/users/elmartinbaron/followers'
-      },
-      profile_image: {
-        small:
-          'https://images.unsplash.com/profile-fb-1599490977-e82172320972.jpg?ixlib=rb-4.0.3&crop=faces&fit=crop&w=32&h=32',
-        medium:
-          'https://images.unsplash.com/profile-fb-1599490977-e82172320972.jpg?ixlib=rb-4.0.3&crop=faces&fit=crop&w=64&h=64',
-        large:
-          'https://images.unsplash.com/profile-fb-1599490977-e82172320972.jpg?ixlib=rb-4.0.3&crop=faces&fit=crop&w=128&h=128'
-      },
-      instagram_username: 'elmartinbaron',
-      total_collections: 0,
-      total_likes: 61,
-      total_photos: 320,
-      total_promoted_photos: 50,
-      accepted_tos: true,
-      for_hire: true,
-      social: {
-        instagram_username: 'elmartinbaron',
-        portfolio_url: 'http://martinbaron.fr',
-        twitter_username: null,
-        paypal_email: null
-      }
-    }
-  },
-  {
-    id: 'Lz2taM_b0CQ',
-    slug: 'a-white-flower-with-green-stems-on-a-white-surface-Lz2taM_b0CQ',
-    created_at: '2024-01-25T16:23:53Z',
-    updated_at: '2024-01-28T14:00:03Z',
-    promoted_at: '2024-01-28T08:44:09Z',
-    width: 2848,
-    height: 4272,
-    color: '#f3f3f3',
-    blur_hash: 'LnK_E?~qjvt701NFxuj]juIUNFt7',
-    description: 'White flower against a beige paper background in a sunny day',
-    alt_description: 'a white flower with green stems on a white surface',
-    breadcrumbs: [],
-    urls: {
-      raw: 'https://images.unsplash.com/photo-1706199405787-90ccfa160fa9?ixid=M3w1NTQ1MTF8MHwxfGFsbHwzfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3',
-      full: 'https://images.unsplash.com/photo-1706199405787-90ccfa160fa9?crop=entropy&cs=srgb&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHwzfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=85',
-      regular:
-        'https://images.unsplash.com/photo-1706199405787-90ccfa160fa9?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHwzfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=1080',
-      small:
-        'https://images.unsplash.com/photo-1706199405787-90ccfa160fa9?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHwzfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=400',
-      thumb:
-        'https://images.unsplash.com/photo-1706199405787-90ccfa160fa9?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHwzfHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=200',
-      small_s3:
-        'https://s3.us-west-2.amazonaws.com/images.unsplash.com/small/photo-1706199405787-90ccfa160fa9'
-    },
-    links: {
-      self: 'https://api.unsplash.com/photos/a-white-flower-with-green-stems-on-a-white-surface-Lz2taM_b0CQ',
-      html: 'https://unsplash.com/photos/a-white-flower-with-green-stems-on-a-white-surface-Lz2taM_b0CQ',
-      download:
-        'https://unsplash.com/photos/Lz2taM_b0CQ/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHwzfHx8fHx8Mnx8MTcwNjQ1OTYxOXw',
-      download_location:
-        'https://api.unsplash.com/photos/Lz2taM_b0CQ/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHwzfHx8fHx8Mnx8MTcwNjQ1OTYxOXw'
-    },
-    likes: 37,
-    liked_by_user: false,
-    current_user_collections: [],
-    sponsorship: null,
-    topic_submissions: {},
-    user: {
-      id: 'VCA9IfCfNMo',
-      updated_at: '2024-01-28T08:47:08Z',
-      username: 'responsiblecreatives',
-      name: 'Keila Hötzel',
-      first_name: 'Keila',
-      last_name: 'Hötzel',
-      twitter_username: 'keilahoetzel',
-      portfolio_url: 'https://adobe.ly/3OJf1GB',
-      bio: 'The german half of The Responsible Creatives, who enjoys shooting styled images and helping brands create their unique identity!  For more images visit our AdobeStock profile. ',
-      location: 'Bayern, Germany',
-      links: {
-        self: 'https://api.unsplash.com/users/responsiblecreatives',
-        html: 'https://unsplash.com/@responsiblecreatives',
-        photos: 'https://api.unsplash.com/users/responsiblecreatives/photos',
-        likes: 'https://api.unsplash.com/users/responsiblecreatives/likes',
-        portfolio:
-          'https://api.unsplash.com/users/responsiblecreatives/portfolio',
-        following:
-          'https://api.unsplash.com/users/responsiblecreatives/following',
-        followers:
-          'https://api.unsplash.com/users/responsiblecreatives/followers'
-      },
-      profile_image: {
-        small:
-          'https://images.unsplash.com/profile-1672753754051-1424c98f5935image?ixlib=rb-4.0.3&crop=faces&fit=crop&w=32&h=32',
-        medium:
-          'https://images.unsplash.com/profile-1672753754051-1424c98f5935image?ixlib=rb-4.0.3&crop=faces&fit=crop&w=64&h=64',
-        large:
-          'https://images.unsplash.com/profile-1672753754051-1424c98f5935image?ixlib=rb-4.0.3&crop=faces&fit=crop&w=128&h=128'
-      },
-      instagram_username: 'responsiblecreatives',
-      total_collections: 2,
-      total_likes: 383,
-      total_photos: 33,
-      total_promoted_photos: 12,
-      accepted_tos: true,
-      for_hire: true,
-      social: {
-        instagram_username: 'responsiblecreatives',
-        portfolio_url: 'https://adobe.ly/3OJf1GB',
-        twitter_username: 'keilahoetzel',
-        paypal_email: null
-      }
-    }
-  },
-  {
-    id: 'wwEWavpRcdk',
-    slug: 'a-green-car-with-a-small-tree-in-the-back-wwEWavpRcdk',
-    created_at: '2024-01-26T20:10:44Z',
-    updated_at: '2024-01-28T16:08:57Z',
-    promoted_at: '2024-01-28T08:43:48Z',
-    width: 4000,
-    height: 6000,
-    color: '#0c260c',
-    blur_hash: 'LnGIiiD%ofRj~pR*RjbH?HazWCoe',
-    description: null,
-    alt_description: 'a green car with a small tree in the back',
-    breadcrumbs: [],
-    urls: {
-      raw: 'https://images.unsplash.com/photo-1706299573468-eafc2096fe94?ixid=M3w1NTQ1MTF8MHwxfGFsbHw0fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3',
-      full: 'https://images.unsplash.com/photo-1706299573468-eafc2096fe94?crop=entropy&cs=srgb&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw0fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=85',
-      regular:
-        'https://images.unsplash.com/photo-1706299573468-eafc2096fe94?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw0fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=1080',
-      small:
-        'https://images.unsplash.com/photo-1706299573468-eafc2096fe94?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw0fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=400',
-      thumb:
-        'https://images.unsplash.com/photo-1706299573468-eafc2096fe94?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw0fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=200',
-      small_s3:
-        'https://s3.us-west-2.amazonaws.com/images.unsplash.com/small/photo-1706299573468-eafc2096fe94'
-    },
-    links: {
-      self: 'https://api.unsplash.com/photos/a-green-car-with-a-small-tree-in-the-back-wwEWavpRcdk',
-      html: 'https://unsplash.com/photos/a-green-car-with-a-small-tree-in-the-back-wwEWavpRcdk',
-      download:
-        'https://unsplash.com/photos/wwEWavpRcdk/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHw0fHx8fHx8Mnx8MTcwNjQ1OTYxOXw',
-      download_location:
-        'https://api.unsplash.com/photos/wwEWavpRcdk/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHw0fHx8fHx8Mnx8MTcwNjQ1OTYxOXw'
-    },
-    likes: 19,
-    liked_by_user: false,
-    current_user_collections: [],
-    sponsorship: null,
-    topic_submissions: {},
-    user: {
-      id: '-Wn6bsBqrws',
-      updated_at: '2024-01-28T08:43:48Z',
-      username: 'karpic',
-      name: 'David Karp.',
-      first_name: 'David',
-      last_name: 'Karp.',
-      twitter_username: null,
-      portfolio_url: 'https://www.instagram.com/karp.pic/',
-      bio: 'Dude with a camera)',
-      location: 'Lviv 📍',
-      links: {
-        self: 'https://api.unsplash.com/users/karpic',
-        html: 'https://unsplash.com/@karpic',
-        photos: 'https://api.unsplash.com/users/karpic/photos',
-        likes: 'https://api.unsplash.com/users/karpic/likes',
-        portfolio: 'https://api.unsplash.com/users/karpic/portfolio',
-        following: 'https://api.unsplash.com/users/karpic/following',
-        followers: 'https://api.unsplash.com/users/karpic/followers'
-      },
-      profile_image: {
-        small:
-          'https://images.unsplash.com/profile-1680818145678-9ed2bad44eae?ixlib=rb-4.0.3&crop=faces&fit=crop&w=32&h=32',
-        medium:
-          'https://images.unsplash.com/profile-1680818145678-9ed2bad44eae?ixlib=rb-4.0.3&crop=faces&fit=crop&w=64&h=64',
-        large:
-          'https://images.unsplash.com/profile-1680818145678-9ed2bad44eae?ixlib=rb-4.0.3&crop=faces&fit=crop&w=128&h=128'
-      },
-      instagram_username: 'https://www.instagram.com/karp.pic/',
-      total_collections: 0,
-      total_likes: 54,
-      total_photos: 88,
-      total_promoted_photos: 11,
-      accepted_tos: true,
-      for_hire: true,
-      social: {
-        instagram_username: 'https://www.instagram.com/karp.pic/',
-        portfolio_url: 'https://www.instagram.com/karp.pic/',
-        twitter_username: null,
-        paypal_email: null
-      }
-    }
-  },
-  {
-    id: 'vYhPMtsQtOI',
-    slug: 'a-small-waterfall-in-the-middle-of-a-forest-vYhPMtsQtOI',
-    created_at: '2024-01-27T00:00:51Z',
-    updated_at: '2024-01-28T08:42:41Z',
-    promoted_at: '2024-01-28T08:42:41Z',
-    width: 5464,
-    height: 6830,
-    color: '#26260c',
-    blur_hash: 'L67nUS-;4TjI-pxu-;?vxb%M?b%2',
-    description: null,
-    alt_description: 'a small waterfall in the middle of a forest',
-    breadcrumbs: [],
-    urls: {
-      raw: 'https://images.unsplash.com/photo-1706313293815-545e6da761be?ixid=M3w1NTQ1MTF8MHwxfGFsbHw1fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3',
-      full: 'https://images.unsplash.com/photo-1706313293815-545e6da761be?crop=entropy&cs=srgb&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw1fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=85',
-      regular:
-        'https://images.unsplash.com/photo-1706313293815-545e6da761be?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw1fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=1080',
-      small:
-        'https://images.unsplash.com/photo-1706313293815-545e6da761be?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw1fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=400',
-      thumb:
-        'https://images.unsplash.com/photo-1706313293815-545e6da761be?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw1fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=200',
-      small_s3:
-        'https://s3.us-west-2.amazonaws.com/images.unsplash.com/small/photo-1706313293815-545e6da761be'
-    },
-    links: {
-      self: 'https://api.unsplash.com/photos/a-small-waterfall-in-the-middle-of-a-forest-vYhPMtsQtOI',
-      html: 'https://unsplash.com/photos/a-small-waterfall-in-the-middle-of-a-forest-vYhPMtsQtOI',
-      download:
-        'https://unsplash.com/photos/vYhPMtsQtOI/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHw1fHx8fHx8Mnx8MTcwNjQ1OTYxOXw',
-      download_location:
-        'https://api.unsplash.com/photos/vYhPMtsQtOI/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHw1fHx8fHx8Mnx8MTcwNjQ1OTYxOXw'
-    },
-    likes: 30,
-    liked_by_user: false,
-    current_user_collections: [],
-    sponsorship: null,
-    topic_submissions: {},
-    user: {
-      id: 'SJ2_XzVF3D0',
-      updated_at: '2024-01-28T16:32:20Z',
-      username: 'ericmuhr',
-      name: 'Eric Muhr',
-      first_name: 'Eric',
-      last_name: 'Muhr',
-      twitter_username: 'ericmuhr',
-      portfolio_url: 'https://linktr.ee/ericmuhr',
-      bio: 'PNW fine art photography | he/him | Prints and originals available | Founder of @oregonexplored',
-      location: 'Portland, OR ',
-      links: {
-        self: 'https://api.unsplash.com/users/ericmuhr',
-        html: 'https://unsplash.com/@ericmuhr',
-        photos: 'https://api.unsplash.com/users/ericmuhr/photos',
-        likes: 'https://api.unsplash.com/users/ericmuhr/likes',
-        portfolio: 'https://api.unsplash.com/users/ericmuhr/portfolio',
-        following: 'https://api.unsplash.com/users/ericmuhr/following',
-        followers: 'https://api.unsplash.com/users/ericmuhr/followers'
-      },
-      profile_image: {
-        small:
-          'https://images.unsplash.com/profile-1617394472428-308356931ad4image?ixlib=rb-4.0.3&crop=faces&fit=crop&w=32&h=32',
-        medium:
-          'https://images.unsplash.com/profile-1617394472428-308356931ad4image?ixlib=rb-4.0.3&crop=faces&fit=crop&w=64&h=64',
-        large:
-          'https://images.unsplash.com/profile-1617394472428-308356931ad4image?ixlib=rb-4.0.3&crop=faces&fit=crop&w=128&h=128'
-      },
-      instagram_username: 'ericmuhr',
-      total_collections: 15,
-      total_likes: 1587,
-      total_photos: 411,
-      total_promoted_photos: 223,
-      accepted_tos: true,
-      for_hire: true,
-      social: {
-        instagram_username: 'ericmuhr',
-        portfolio_url: 'https://linktr.ee/ericmuhr',
-        twitter_username: 'ericmuhr',
-        paypal_email: null
-      }
-    }
-  },
-  {
-    id: 'igw-14UnNjg',
-    slug: 'a-couple-of-people-that-are-walking-in-the-dirt-igw-14UnNjg',
-    created_at: '2023-04-28T15:30:26Z',
-    updated_at: '2024-01-22T00:15:38Z',
-    promoted_at: '2023-05-17T09:52:47Z',
-    width: 9504,
-    height: 6336,
-    color: '#c0a68c',
-    blur_hash: 'LLJ@dv4oN{%M0g9aoI-o}=NGIUx]',
-    description:
-      'A beautiful and ancient spring fed canyon which weaves its way through 400-meter-tall towers of granite, sandstone and basalt, before plunging into the Gulf of Aqaba | Shiʻb Mūsá – NEOM, Saudi Arabia.',
-    alt_description: 'a couple of people that are walking in the dirt',
-    breadcrumbs: [],
-    urls: {
-      raw: 'https://images.unsplash.com/photo-1682695795557-17447f921f79?ixid=M3w1NTQ1MTF8MXwxfGFsbHw2fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3',
-      full: 'https://images.unsplash.com/photo-1682695795557-17447f921f79?crop=entropy&cs=srgb&fm=jpg&ixid=M3w1NTQ1MTF8MXwxfGFsbHw2fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=85',
-      regular:
-        'https://images.unsplash.com/photo-1682695795557-17447f921f79?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MXwxfGFsbHw2fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=1080',
-      small:
-        'https://images.unsplash.com/photo-1682695795557-17447f921f79?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MXwxfGFsbHw2fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=400',
-      thumb:
-        'https://images.unsplash.com/photo-1682695795557-17447f921f79?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MXwxfGFsbHw2fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=200',
-      small_s3:
-        'https://s3.us-west-2.amazonaws.com/images.unsplash.com/small/photo-1682695795557-17447f921f79'
-    },
-    links: {
-      self: 'https://api.unsplash.com/photos/a-couple-of-people-that-are-walking-in-the-dirt-igw-14UnNjg',
-      html: 'https://unsplash.com/photos/a-couple-of-people-that-are-walking-in-the-dirt-igw-14UnNjg',
-      download:
-        'https://unsplash.com/photos/igw-14UnNjg/download?ixid=M3w1NTQ1MTF8MXwxfGFsbHw2fHx8fHx8Mnx8MTcwNjQ1OTYxOXw',
-      download_location:
-        'https://api.unsplash.com/photos/igw-14UnNjg/download?ixid=M3w1NTQ1MTF8MXwxfGFsbHw2fHx8fHx8Mnx8MTcwNjQ1OTYxOXw'
-    },
-    likes: 250,
-    liked_by_user: false,
-    current_user_collections: [],
-    sponsorship: {
-      impression_urls: [
-        'https://secure.insightexpressai.com/adServer/adServerESI.aspx?script=false&bannerID=11515591&rnd=[timestamp]&redir=https://secure.insightexpressai.com/adserver/1pixel.gif',
-        'https://secure.insightexpressai.com/adServer/adServerESI.aspx?script=false&bannerID=11515794&rnd=[timestamp]&redir=https://secure.insightexpressai.com/adserver/1pixel.gif'
-      ],
-      tagline: 'Made to Change',
-      tagline_url:
-        'https://www.neom.com/en-us?utm_source=unsplash&utm_medium=referral',
-      sponsor: {
-        id: 'mYizSrdJkkU',
-        updated_at: '2024-01-27T21:36:45Z',
-        username: 'neom',
-        name: 'NEOM',
-        first_name: 'NEOM',
-        last_name: null,
-        twitter_username: 'neom',
-        portfolio_url: 'http://www.neom.com',
-        bio: 'Located in the northwest of Saudi Arabia, NEOM’s diverse climate offers both sun-soaked beaches and snow-capped mountains. NEOM’s unique location will provide residents with enhanced livability while protecting 95% of the natural landscape.',
-        location: 'NEOM, Saudi Arabia',
-        links: {
-          self: 'https://api.unsplash.com/users/neom',
-          html: 'https://unsplash.com/@neom',
-          photos: 'https://api.unsplash.com/users/neom/photos',
-          likes: 'https://api.unsplash.com/users/neom/likes',
-          portfolio: 'https://api.unsplash.com/users/neom/portfolio',
-          following: 'https://api.unsplash.com/users/neom/following',
-          followers: 'https://api.unsplash.com/users/neom/followers'
-        },
-        profile_image: {
-          small:
-            'https://images.unsplash.com/profile-1679489218992-ebe823c797dfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=32&h=32',
-          medium:
-            'https://images.unsplash.com/profile-1679489218992-ebe823c797dfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=64&h=64',
-          large:
-            'https://images.unsplash.com/profile-1679489218992-ebe823c797dfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=128&h=128'
-        },
-        instagram_username: 'discoverneom',
-        total_collections: 7,
-        total_likes: 1,
-        total_photos: 202,
-        total_promoted_photos: 72,
-        accepted_tos: true,
-        for_hire: false,
-        social: {
-          instagram_username: 'discoverneom',
-          portfolio_url: 'http://www.neom.com',
-          twitter_username: 'neom',
-          paypal_email: null
-        }
-      }
-    },
-    topic_submissions: {},
-    user: {
-      id: 'mYizSrdJkkU',
-      updated_at: '2024-01-27T21:36:45Z',
-      username: 'neom',
-      name: 'NEOM',
-      first_name: 'NEOM',
-      last_name: null,
-      twitter_username: 'neom',
-      portfolio_url: 'http://www.neom.com',
-      bio: 'Located in the northwest of Saudi Arabia, NEOM’s diverse climate offers both sun-soaked beaches and snow-capped mountains. NEOM’s unique location will provide residents with enhanced livability while protecting 95% of the natural landscape.',
-      location: 'NEOM, Saudi Arabia',
-      links: {
-        self: 'https://api.unsplash.com/users/neom',
-        html: 'https://unsplash.com/@neom',
-        photos: 'https://api.unsplash.com/users/neom/photos',
-        likes: 'https://api.unsplash.com/users/neom/likes',
-        portfolio: 'https://api.unsplash.com/users/neom/portfolio',
-        following: 'https://api.unsplash.com/users/neom/following',
-        followers: 'https://api.unsplash.com/users/neom/followers'
-      },
-      profile_image: {
-        small:
-          'https://images.unsplash.com/profile-1679489218992-ebe823c797dfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=32&h=32',
-        medium:
-          'https://images.unsplash.com/profile-1679489218992-ebe823c797dfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=64&h=64',
-        large:
-          'https://images.unsplash.com/profile-1679489218992-ebe823c797dfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=128&h=128'
-      },
-      instagram_username: 'discoverneom',
-      total_collections: 7,
-      total_likes: 1,
-      total_photos: 202,
-      total_promoted_photos: 72,
-      accepted_tos: true,
-      for_hire: false,
-      social: {
-        instagram_username: 'discoverneom',
-        portfolio_url: 'http://www.neom.com',
-        twitter_username: 'neom',
-        paypal_email: null
-      }
-    }
-  },
-  {
-    id: 'rlLWrMZ4Qx8',
-    slug: 'a-woman-in-a-straw-hat-holding-a-stick-rlLWrMZ4Qx8',
-    created_at: '2024-01-27T11:29:41Z',
-    updated_at: '2024-01-28T08:42:37Z',
-    promoted_at: '2024-01-28T08:42:37Z',
-    width: 2200,
-    height: 2750,
-    color: '#26260c',
-    blur_hash: 'LAA0:V4p0M~n?YWCbbV@4:x@-:D+',
-    description: null,
-    alt_description: 'a woman in a straw hat holding a stick',
-    breadcrumbs: [],
-    urls: {
-      raw: 'https://images.unsplash.com/photo-1706354924674-0304751469e8?ixid=M3w1NTQ1MTF8MHwxfGFsbHw3fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3',
-      full: 'https://images.unsplash.com/photo-1706354924674-0304751469e8?crop=entropy&cs=srgb&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw3fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=85',
-      regular:
-        'https://images.unsplash.com/photo-1706354924674-0304751469e8?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw3fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=1080',
-      small:
-        'https://images.unsplash.com/photo-1706354924674-0304751469e8?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw3fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=400',
-      thumb:
-        'https://images.unsplash.com/photo-1706354924674-0304751469e8?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw3fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=200',
-      small_s3:
-        'https://s3.us-west-2.amazonaws.com/images.unsplash.com/small/photo-1706354924674-0304751469e8'
-    },
-    links: {
-      self: 'https://api.unsplash.com/photos/a-woman-in-a-straw-hat-holding-a-stick-rlLWrMZ4Qx8',
-      html: 'https://unsplash.com/photos/a-woman-in-a-straw-hat-holding-a-stick-rlLWrMZ4Qx8',
-      download:
-        'https://unsplash.com/photos/rlLWrMZ4Qx8/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHw3fHx8fHx8Mnx8MTcwNjQ1OTYxOXw',
-      download_location:
-        'https://api.unsplash.com/photos/rlLWrMZ4Qx8/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHw3fHx8fHx8Mnx8MTcwNjQ1OTYxOXw'
-    },
-    likes: 17,
-    liked_by_user: false,
-    current_user_collections: [],
-    sponsorship: null,
-    topic_submissions: {},
-    user: {
-      id: 'aqGpsQ3Jl5E',
-      updated_at: '2024-01-28T08:42:37Z',
-      username: 'mahdi_chf',
-      name: 'mahdi chaghari',
-      first_name: 'mahdi',
-      last_name: 'chaghari',
-      twitter_username: 'mahdi_chf',
-      portfolio_url: 'http://instagram.com/mahdi__chf',
-      bio: '"good mistake"',
-      location: null,
-      links: {
-        self: 'https://api.unsplash.com/users/mahdi_chf',
-        html: 'https://unsplash.com/@mahdi_chf',
-        photos: 'https://api.unsplash.com/users/mahdi_chf/photos',
-        likes: 'https://api.unsplash.com/users/mahdi_chf/likes',
-        portfolio: 'https://api.unsplash.com/users/mahdi_chf/portfolio',
-        following: 'https://api.unsplash.com/users/mahdi_chf/following',
-        followers: 'https://api.unsplash.com/users/mahdi_chf/followers'
-      },
-      profile_image: {
-        small:
-          'https://images.unsplash.com/profile-1609792897842-d72d944d33cfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=32&h=32',
-        medium:
-          'https://images.unsplash.com/profile-1609792897842-d72d944d33cfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=64&h=64',
-        large:
-          'https://images.unsplash.com/profile-1609792897842-d72d944d33cfimage?ixlib=rb-4.0.3&crop=faces&fit=crop&w=128&h=128'
-      },
-      instagram_username: 'mahdi_chf',
-      total_collections: 0,
-      total_likes: 45,
-      total_photos: 652,
-      total_promoted_photos: 39,
-      accepted_tos: true,
-      for_hire: false,
-      social: {
-        instagram_username: 'mahdi_chf',
-        portfolio_url: 'http://instagram.com/mahdi__chf',
-        twitter_username: 'mahdi_chf',
-        paypal_email: null
-      }
-    }
-  },
-  {
-    id: '9duuU4kMI6s',
-    slug: 'a-close-up-of-a-sunflower-with-a-blurry-background-9duuU4kMI6s',
-    created_at: '2024-01-23T11:35:32Z',
-    updated_at: '2024-01-28T10:56:18Z',
-    promoted_at: '2024-01-28T08:42:32Z',
-    width: 6016,
-    height: 4016,
-    color: '#0c260c',
-    blur_hash: 'L76ahLS40jsm-6R-I[xWIrRlt5xr',
-    description: null,
-    alt_description: 'a close up of a sunflower with a blurry background',
-    breadcrumbs: [],
-    urls: {
-      raw: 'https://images.unsplash.com/photo-1706007171907-49841a64792a?ixid=M3w1NTQ1MTF8MHwxfGFsbHw4fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3',
-      full: 'https://images.unsplash.com/photo-1706007171907-49841a64792a?crop=entropy&cs=srgb&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw4fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=85',
-      regular:
-        'https://images.unsplash.com/photo-1706007171907-49841a64792a?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw4fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=1080',
-      small:
-        'https://images.unsplash.com/photo-1706007171907-49841a64792a?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw4fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=400',
-      thumb:
-        'https://images.unsplash.com/photo-1706007171907-49841a64792a?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw4fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=200',
-      small_s3:
-        'https://s3.us-west-2.amazonaws.com/images.unsplash.com/small/photo-1706007171907-49841a64792a'
-    },
-    links: {
-      self: 'https://api.unsplash.com/photos/a-close-up-of-a-sunflower-with-a-blurry-background-9duuU4kMI6s',
-      html: 'https://unsplash.com/photos/a-close-up-of-a-sunflower-with-a-blurry-background-9duuU4kMI6s',
-      download:
-        'https://unsplash.com/photos/9duuU4kMI6s/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHw4fHx8fHx8Mnx8MTcwNjQ1OTYxOXw',
-      download_location:
-        'https://api.unsplash.com/photos/9duuU4kMI6s/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHw4fHx8fHx8Mnx8MTcwNjQ1OTYxOXw'
-    },
-    likes: 16,
-    liked_by_user: false,
-    current_user_collections: [],
-    sponsorship: null,
-    topic_submissions: {},
-    user: {
-      id: 'tBplnoPr52M',
-      updated_at: '2024-01-28T08:42:32Z',
-      username: 'cherstve_pechivo',
-      name: 'Liana S',
-      first_name: 'Liana',
-      last_name: 'S',
-      twitter_username: null,
-      portfolio_url: null,
-      bio: 'Hello! I love the beauty of this world and want to share it with you. Look',
-      location: 'Kyiv, Ukraine',
-      links: {
-        self: 'https://api.unsplash.com/users/cherstve_pechivo',
-        html: 'https://unsplash.com/@cherstve_pechivo',
-        photos: 'https://api.unsplash.com/users/cherstve_pechivo/photos',
-        likes: 'https://api.unsplash.com/users/cherstve_pechivo/likes',
-        portfolio: 'https://api.unsplash.com/users/cherstve_pechivo/portfolio',
-        following: 'https://api.unsplash.com/users/cherstve_pechivo/following',
-        followers: 'https://api.unsplash.com/users/cherstve_pechivo/followers'
-      },
-      profile_image: {
-        small:
-          'https://images.unsplash.com/profile-fb-1705742124-9ed5f93ab36e.jpg?ixlib=rb-4.0.3&crop=faces&fit=crop&w=32&h=32',
-        medium:
-          'https://images.unsplash.com/profile-fb-1705742124-9ed5f93ab36e.jpg?ixlib=rb-4.0.3&crop=faces&fit=crop&w=64&h=64',
-        large:
-          'https://images.unsplash.com/profile-fb-1705742124-9ed5f93ab36e.jpg?ixlib=rb-4.0.3&crop=faces&fit=crop&w=128&h=128'
-      },
-      instagram_username: 'cherstve_pechivo',
-      total_collections: 8,
-      total_likes: 189,
-      total_photos: 56,
-      total_promoted_photos: 2,
-      accepted_tos: true,
-      for_hire: false,
-      social: {
-        instagram_username: 'cherstve_pechivo',
-        portfolio_url: null,
-        twitter_username: null,
-        paypal_email: null
-      }
-    }
-  },
-  {
-    id: '8rQQBxAtvL4',
-    slug: 'a-painting-of-a-colorful-bird-on-a-white-background-8rQQBxAtvL4',
-    created_at: '2024-01-26T11:06:45Z',
-    updated_at: '2024-01-28T09:11:54Z',
-    promoted_at: '2024-01-28T08:40:56Z',
-    width: 4370,
-    height: 5669,
-    color: '#a6260c',
-    blur_hash: 'LLIgNAIaqZtTRhk=XUadvgX6XnSi',
-    description:
-      "' Urban Light' - a large abstract acryl painting on canvas with a lot of yellow, red and blue colors, which I painted in 2008 together with my artist friend Ben Vollers. Only after the common painting proces we named our work in words. This colorful dynamic painting is one of the c. 50 large paintings we created together: jamming with brushes and paint! Some call this fusion-art or jamming. This free art image is suitable for making large art prints, posters or wallpapers - Colorful art image in public domain CCO - free download, by Fons Heijnsbroek, Dutch painter artist of The Netherlands.",
-    alt_description: 'a painting of a colorful bird on a white background',
-    breadcrumbs: [],
-    urls: {
-      raw: 'https://images.unsplash.com/photo-1706266541244-89bfdd5ad021?ixid=M3w1NTQ1MTF8MHwxfGFsbHw5fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3',
-      full: 'https://images.unsplash.com/photo-1706266541244-89bfdd5ad021?crop=entropy&cs=srgb&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw5fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=85',
-      regular:
-        'https://images.unsplash.com/photo-1706266541244-89bfdd5ad021?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw5fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=1080',
-      small:
-        'https://images.unsplash.com/photo-1706266541244-89bfdd5ad021?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw5fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=400',
-      thumb:
-        'https://images.unsplash.com/photo-1706266541244-89bfdd5ad021?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHw5fHx8fHx8Mnx8MTcwNjQ1OTYxOXw&ixlib=rb-4.0.3&q=80&w=200',
-      small_s3:
-        'https://s3.us-west-2.amazonaws.com/images.unsplash.com/small/photo-1706266541244-89bfdd5ad021'
-    },
-    links: {
-      self: 'https://api.unsplash.com/photos/a-painting-of-a-colorful-bird-on-a-white-background-8rQQBxAtvL4',
-      html: 'https://unsplash.com/photos/a-painting-of-a-colorful-bird-on-a-white-background-8rQQBxAtvL4',
-      download:
-        'https://unsplash.com/photos/8rQQBxAtvL4/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHw5fHx8fHx8Mnx8MTcwNjQ1OTYxOXw',
-      download_location:
-        'https://api.unsplash.com/photos/8rQQBxAtvL4/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHw5fHx8fHx8Mnx8MTcwNjQ1OTYxOXw'
-    },
-    likes: 21,
-    liked_by_user: false,
-    current_user_collections: [],
-    sponsorship: null,
-    topic_submissions: {
-      'textures-patterns': {
-        status: 'unevaluated'
-      },
-      wallpapers: {
-        status: 'unevaluated'
-      }
-    },
-    user: {
-      id: 'lGsjSEeF3XY',
-      updated_at: '2024-01-28T16:22:19Z',
-      username: 'heijnsbroek_abstract_art',
-      name: 'Fons Heijnsbroek, abstract-art',
-      first_name: 'Fons',
-      last_name: 'Heijnsbroek, abstract-art',
-      twitter_username: null,
-      portfolio_url: 'https://heijnsbroek.exto.org/',
-      bio: 'My abstract colorful paintings I made in 1985 - 2016. Here I share photos of my mainly abstract art in high resolution pictures, suitable for art print or wallpaper - free download & public domain art images CCO. I appreciate when you mention my name',
-      location: 'Amsterdam, Netherlands, Holland',
-      links: {
-        self: 'https://api.unsplash.com/users/heijnsbroek_abstract_art',
-        html: 'https://unsplash.com/@heijnsbroek_abstract_art',
-        photos:
-          'https://api.unsplash.com/users/heijnsbroek_abstract_art/photos',
-        likes: 'https://api.unsplash.com/users/heijnsbroek_abstract_art/likes',
-        portfolio:
-          'https://api.unsplash.com/users/heijnsbroek_abstract_art/portfolio',
-        following:
-          'https://api.unsplash.com/users/heijnsbroek_abstract_art/following',
-        followers:
-          'https://api.unsplash.com/users/heijnsbroek_abstract_art/followers'
-      },
-      profile_image: {
-        small:
-          'https://images.unsplash.com/profile-1704709670831-27061c696054image?ixlib=rb-4.0.3&crop=faces&fit=crop&w=32&h=32',
-        medium:
-          'https://images.unsplash.com/profile-1704709670831-27061c696054image?ixlib=rb-4.0.3&crop=faces&fit=crop&w=64&h=64',
-        large:
-          'https://images.unsplash.com/profile-1704709670831-27061c696054image?ixlib=rb-4.0.3&crop=faces&fit=crop&w=128&h=128'
-      },
-      instagram_username: 'heijnsbroekfons',
-      total_collections: 4,
-      total_likes: 420,
-      total_photos: 88,
-      total_promoted_photos: 7,
-      accepted_tos: true,
-      for_hire: false,
-      social: {
-        instagram_username: 'heijnsbroekfons',
-        portfolio_url: 'https://heijnsbroek.exto.org/',
-        twitter_username: null,
-        paypal_email: null
-      }
-    }
-  },
-  {
-    id: 'v4j0rlrTZbc',
-    slug: 'a-mountain-covered-in-snow-at-sunset-v4j0rlrTZbc',
-    created_at: '2024-01-24T10:38:01Z',
-    updated_at: '2024-01-28T14:00:03Z',
-    promoted_at: '2024-01-28T08:40:50Z',
-    width: 5184,
-    height: 3456,
-    color: '#0c2626',
-    blur_hash: 'L^J?{G%0s.WV~Bs:j[azOZR-WXoe',
-    description:
-      'Sunrise hike across Bali to view the beautiful Mt. Batur in Bali, Indonesia.',
-    alt_description: 'a mountain covered in snow at sunset',
-    breadcrumbs: [],
-    urls: {
-      raw: 'https://images.unsplash.com/photo-1706088604350-f3c46951ee02?ixid=M3w1NTQ1MTF8MHwxfGFsbHwxMHx8fHx8fDJ8fDE3MDY0NTk2MTl8&ixlib=rb-4.0.3',
-      full: 'https://images.unsplash.com/photo-1706088604350-f3c46951ee02?crop=entropy&cs=srgb&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHwxMHx8fHx8fDJ8fDE3MDY0NTk2MTl8&ixlib=rb-4.0.3&q=85',
-      regular:
-        'https://images.unsplash.com/photo-1706088604350-f3c46951ee02?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHwxMHx8fHx8fDJ8fDE3MDY0NTk2MTl8&ixlib=rb-4.0.3&q=80&w=1080',
-      small:
-        'https://images.unsplash.com/photo-1706088604350-f3c46951ee02?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHwxMHx8fHx8fDJ8fDE3MDY0NTk2MTl8&ixlib=rb-4.0.3&q=80&w=400',
-      thumb:
-        'https://images.unsplash.com/photo-1706088604350-f3c46951ee02?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=M3w1NTQ1MTF8MHwxfGFsbHwxMHx8fHx8fDJ8fDE3MDY0NTk2MTl8&ixlib=rb-4.0.3&q=80&w=200',
-      small_s3:
-        'https://s3.us-west-2.amazonaws.com/images.unsplash.com/small/photo-1706088604350-f3c46951ee02'
-    },
-    links: {
-      self: 'https://api.unsplash.com/photos/a-mountain-covered-in-snow-at-sunset-v4j0rlrTZbc',
-      html: 'https://unsplash.com/photos/a-mountain-covered-in-snow-at-sunset-v4j0rlrTZbc',
-      download:
-        'https://unsplash.com/photos/v4j0rlrTZbc/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHwxMHx8fHx8fDJ8fDE3MDY0NTk2MTl8',
-      download_location:
-        'https://api.unsplash.com/photos/v4j0rlrTZbc/download?ixid=M3w1NTQ1MTF8MHwxfGFsbHwxMHx8fHx8fDJ8fDE3MDY0NTk2MTl8'
-    },
-    likes: 24,
-    liked_by_user: false,
-    current_user_collections: [],
-    sponsorship: null,
-    topic_submissions: {},
-    user: {
-      id: '-Mt89NUTYp0',
-      updated_at: '2024-01-28T09:02:08Z',
-      username: 'carltraw',
-      name: 'Carl Raw',
-      first_name: 'Carl',
-      last_name: 'Raw',
-      twitter_username: 'carltraw',
-      portfolio_url: 'http://www.carlraw.co.uk',
-      bio: "I'm a designer, photographer, content creator and image manipulator.",
-      location: 'Liverpool',
-      links: {
-        self: 'https://api.unsplash.com/users/carltraw',
-        html: 'https://unsplash.com/@carltraw',
-        photos: 'https://api.unsplash.com/users/carltraw/photos',
-        likes: 'https://api.unsplash.com/users/carltraw/likes',
-        portfolio: 'https://api.unsplash.com/users/carltraw/portfolio',
-        following: 'https://api.unsplash.com/users/carltraw/following',
-        followers: 'https://api.unsplash.com/users/carltraw/followers'
-      },
-      profile_image: {
-        small:
-          'https://images.unsplash.com/profile-1659961233253-da31f215e037image?ixlib=rb-4.0.3&crop=faces&fit=crop&w=32&h=32',
-        medium:
-          'https://images.unsplash.com/profile-1659961233253-da31f215e037image?ixlib=rb-4.0.3&crop=faces&fit=crop&w=64&h=64',
-        large:
-          'https://images.unsplash.com/profile-1659961233253-da31f215e037image?ixlib=rb-4.0.3&crop=faces&fit=crop&w=128&h=128'
-      },
-      instagram_username: 'rawimage.co.uk',
-      total_collections: 0,
-      total_likes: 238,
-      total_photos: 65,
-      total_promoted_photos: 19,
-      accepted_tos: true,
-      for_hire: true,
-      social: {
-        instagram_username: 'rawimage.co.uk',
-        portfolio_url: 'http://www.carlraw.co.uk',
-        twitter_username: 'carltraw',
-        paypal_email: null
-      }
-    }
-  }
-];
-
-export function PhotoSection() {
-  return (
-    <section className={s.section}>
-      <ul className={s.photoList}>
-        {photoList.map((photoInfo: PhotoInfo) => (
-          <li
-            key={photoInfo.id}
-            className={clsx(s.photoListItem, {
-              [s.photoListItemLarge]: photoInfo.width < photoInfo.height
-            })}
-          >
-            <figure className={s.figure}>
-              <img
-                className={s.photoImg}
-                src={photoInfo.urls.raw}
-                alt={photoInfo.alt_description}
-              />
-              <figcaption className={s.figcaption}>
-                <Link
-                  to="/aston-react-project/photos/:id"
-                  className={s.infoButton}
-                >
-                  Больше информации
-                </Link>
-                <LikeButton />
-              </figcaption>
-            </figure>
-          </li>
-        ))}
-      </ul>
-    </section>
-  );
-}
