+++ src/shared/ui/Section/Section.module.css
@@ -1,0 +1,21 @@
+.topSection {
+  width: 100%;
+  min-height: 100px;
+  background-color: var(--light-theme-accent-color);
+  border-radius: 60px 60px 0px 0px;
+  padding: 91px 0px 60px;
+}
+
+.photoSection {
+  position: relative;
+  max-width: 100%;
+  min-height: 550px;
+  background-color: var(--light-theme-background-color);
+  border-radius: 60px;
+  margin-top: -60px;
+  padding: 60px 0px;
+  box-shadow:
+    0px 14px 80px rgba(0, 0, 0, 0.25),
+    0px -14px 80px rgba(0, 0, 0, 0.25);
+  display: flex;
+}
