+++ src/widgets/PhotoList/PhotoList.module.css
@@ -1,18 +1,4 @@
-.section {
-  position: relative;
-  max-width: 100%;
-  min-height: 550px;
-  background-color: var(--light-theme-background-color);
-  border-radius: 60px;
-  margin-top: -60px;
-  padding: 60px 0px;
-  box-shadow:
-    0px 14px 80px rgba(0, 0, 0, 0.25),
-    0px -14px 80px rgba(0, 0, 0, 0.25);
-  display: flex;
-}
-
-.photoList {
+.list {
   width: 91%;
   margin: 0 auto;
   padding: 0;
@@ -23,15 +9,15 @@
   grid-gap: 10px;
 }
 
-.photoListItem {
+.listItem {
   list-style: none;
   margin: 0;
   overflow: hidden;
   border-radius: 10px;
 }
 
-.photoListItemLarge {
-  composes: photoListItem;
+.listItemLarge {
+  composes: listItem;
   grid-column: span 1;
   grid-row: span 2;
 }
@@ -72,7 +58,7 @@
 .figure:after {
   transform: skew(45deg) translateX(190%);
 }
-.figure:hover > .photoImg {
+.figure:hover > .img {
   opacity: 0.5;
 }
 
@@ -89,7 +75,7 @@
   opacity: 1;
 }
 
-.photoImg {
+.img {
   backface-visibility: hidden;
   width: 100%;
   height: 100%;
